requirements.txt
aiogram==2.19
aiohttp==3.8.1
aioschedule==0.5.2
aiosignal==1.2.0
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
blinker==1.4
Brotli==1.0.9
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.6.0
firebase-admin==5.2.0
Flask==2.0.3
frozenlist==1.3.0
gcloud==0.17.0
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
httplib2==0.20.4
hyperframe==6.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
jws==0.1.3
kaitaistruct==0.9
MarkupSafe==2.1.1
MouseInfo==0.1.3
msgpack==1.0.3
multidict==6.0.2
multithreading==0.2.0
numpy==1.22.3
oauth2client==3.0.0
oauthlib==3.2.0
outcome==1.1.0
pandas==1.4.1
Pillow==9.0.1
platformdirs==2.5.1
proto-plus==1.20.3
protobuf==3.19.4
psycopg2==2.9.3
pyaes==1.6.1
pyasn1==0.4.7
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.4.3
PyGetWindow==0.0.9
pygsheets==2.0.5
PyJWT==2.3.0
PyMsgBox==1.0.9
pyobjc==8.3
pyobjc-core==8.3
pyobjc-framework-Accessibility==8.3
pyobjc-framework-Accounts==8.3
pyobjc-framework-AddressBook==8.3
pyobjc-framework-AdServices==8.3
pyobjc-framework-AdSupport==8.3
pyobjc-framework-AppleScriptKit==8.3
pyobjc-framework-AppleScriptObjC==8.3
pyobjc-framework-ApplicationServices==8.3
pyobjc-framework-AppTrackingTransparency==8.3
pyobjc-framework-AudioVideoBridging==8.3
pyobjc-framework-AuthenticationServices==8.3
pyobjc-framework-AutomaticAssessmentConfiguration==8.3
pyobjc-framework-Automator==8.3
pyobjc-framework-AVFoundation==8.3
pyobjc-framework-AVKit==8.3
pyobjc-framework-BusinessChat==8.3
pyobjc-framework-CalendarStore==8.3
pyobjc-framework-CallKit==8.3
pyobjc-framework-CFNetwork==8.3
pyobjc-framework-ClassKit==8.3
pyobjc-framework-CloudKit==8.3
pyobjc-framework-Cocoa==8.3
pyobjc-framework-Collaboration==8.3
pyobjc-framework-ColorSync==8.3
pyobjc-framework-Contacts==8.3
pyobjc-framework-ContactsUI==8.3
pyobjc-framework-CoreAudio==8.3
pyobjc-framework-CoreAudioKit==8.3
pyobjc-framework-CoreBluetooth==8.3
pyobjc-framework-CoreData==8.3
pyobjc-framework-CoreHaptics==8.3
pyobjc-framework-CoreLocation==8.3
pyobjc-framework-CoreMedia==8.3
pyobjc-framework-CoreMediaIO==8.3
pyobjc-framework-CoreMIDI==8.3
pyobjc-framework-CoreML==8.3
pyobjc-framework-CoreMotion==8.3
pyobjc-framework-CoreServices==8.3
pyobjc-framework-CoreSpotlight==8.3
pyobjc-framework-CoreText==8.3
pyobjc-framework-CoreWLAN==8.3
pyobjc-framework-CryptoTokenKit==8.3
pyobjc-framework-DataDetection==8.3
pyobjc-framework-DeviceCheck==8.3
pyobjc-framework-DictionaryServices==8.3
pyobjc-framework-DiscRecording==8.3
pyobjc-framework-DiscRecordingUI==8.3
pyobjc-framework-DiskArbitration==8.3
pyobjc-framework-DVDPlayback==8.3
pyobjc-framework-EventKit==8.3
pyobjc-framework-ExceptionHandling==8.3
pyobjc-framework-ExecutionPolicy==8.3
pyobjc-framework-ExternalAccessory==8.3
pyobjc-framework-FileProvider==8.3
pyobjc-framework-FileProviderUI==8.3
pyobjc-framework-FinderSync==8.3
pyobjc-framework-FSEvents==8.3
pyobjc-framework-GameCenter==8.3
pyobjc-framework-GameController==8.3
pyobjc-framework-GameKit==8.3
pyobjc-framework-GameplayKit==8.3
pyobjc-framework-ImageCaptureCore==8.3
pyobjc-framework-IMServicePlugIn==8.3
pyobjc-framework-InputMethodKit==8.3
pyobjc-framework-InstallerPlugins==8.3
pyobjc-framework-InstantMessage==8.3
pyobjc-framework-Intents==8.3
pyobjc-framework-IntentsUI==8.3
pyobjc-framework-IOSurface==8.3
pyobjc-framework-iTunesLibrary==8.3
pyobjc-framework-KernelManagement==8.3
pyobjc-framework-LatentSemanticMapping==8.3
pyobjc-framework-LaunchServices==8.3
pyobjc-framework-libdispatch==8.3
pyobjc-framework-LinkPresentation==8.3
pyobjc-framework-LocalAuthentication==8.3
pyobjc-framework-LocalAuthenticationEmbeddedUI==8.3
pyobjc-framework-MailKit==8.3
pyobjc-framework-MapKit==8.3
pyobjc-framework-MediaAccessibility==8.3
pyobjc-framework-MediaLibrary==8.3
pyobjc-framework-MediaPlayer==8.3
pyobjc-framework-MediaToolbox==8.3
pyobjc-framework-Metal==8.3
pyobjc-framework-MetalKit==8.3
pyobjc-framework-MetalPerformanceShaders==8.3
pyobjc-framework-MetalPerformanceShadersGraph==8.3
pyobjc-framework-MetricKit==8.3
pyobjc-framework-MLCompute==8.3
pyobjc-framework-ModelIO==8.3
pyobjc-framework-MultipeerConnectivity==8.3
pyobjc-framework-NaturalLanguage==8.3
pyobjc-framework-NetFS==8.3
pyobjc-framework-Network==8.3
pyobjc-framework-NetworkExtension==8.3
pyobjc-framework-NotificationCenter==8.3
pyobjc-framework-OpenDirectory==8.3
pyobjc-framework-OSAKit==8.3
pyobjc-framework-OSLog==8.3
pyobjc-framework-PassKit==8.3
pyobjc-framework-PencilKit==8.3
pyobjc-framework-Photos==8.3
pyobjc-framework-PhotosUI==8.3
pyobjc-framework-PreferencePanes==8.3
pyobjc-framework-PushKit==8.3
pyobjc-framework-Quartz==8.3
pyobjc-framework-QuickLookThumbnailing==8.3
pyobjc-framework-ReplayKit==8.3
pyobjc-framework-SafariServices==8.3
pyobjc-framework-SceneKit==8.3
pyobjc-framework-ScreenSaver==8.3
pyobjc-framework-ScreenTime==8.3
pyobjc-framework-ScriptingBridge==8.3
pyobjc-framework-SearchKit==8.3
pyobjc-framework-Security==8.3
pyobjc-framework-SecurityFoundation==8.3
pyobjc-framework-SecurityInterface==8.3
pyobjc-framework-ServiceManagement==8.3
pyobjc-framework-ShazamKit==8.3
pyobjc-framework-Social==8.3
pyobjc-framework-SoundAnalysis==8.3
pyobjc-framework-Speech==8.3
pyobjc-framework-SpriteKit==8.3
pyobjc-framework-StoreKit==8.3
pyobjc-framework-SyncServices==8.3
pyobjc-framework-SystemConfiguration==8.3
pyobjc-framework-SystemExtensions==8.3
pyobjc-framework-UniformTypeIdentifiers==8.3
pyobjc-framework-UserNotifications==8.3
pyobjc-framework-UserNotificationsUI==8.3
pyobjc-framework-VideoSubscriberAccount==8.3
pyobjc-framework-VideoToolbox==8.3
pyobjc-framework-Virtualization==8.3
pyobjc-framework-Vision==8.3
pyobjc-framework-WebKit==8.3
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyperclip==1.8.2
Pyrebase==3.0.27
PyRect==0.1.4
PyScreeze==0.1.28
PySocks==1.7.1
pyTelegramBotAPI==4.4.0
python-aalib==0.3.2
python-dateutil==2.8.2
python-jwt==2.0.1
python-telegram==0.15.0
pytweening==1.0.4
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.7.0
rsa==4.0
rubicon-objc==0.4.2
schedule==1.1.0
selenium==4.1.2
selenium-wire==4.6.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
Telethon==1.24.0
tg==0.19.0
trio==0.20.0
trio-websocket==0.9.2
uplink==0.9.7
uritemplate==4.1.1
urllib3==1.26.8
uuid==1.30
virtualenv==20.14.0
Werkzeug==2.0.3
whitenoise==6.0.0
wsproto==1.1.0
yarl==1.7.2
zstandard==0.17.0
